requirements.txt
uvicorn
fastapi
litellm
pandas
voyager==2.0.2
gradio
promptflow
promptflow-tools
numpy
scikit-learn

diagrams
graphviz
graphviz2drawio
crewai

duckduckgo-search
python-dotenv
pyautogen
autogenstudio